Add a note about the Discord's role hierarchies
This is important if you are adding a new instance of Chill Bot to a server since the bot's role will be added at the very bottom of the hierarchy by default. This means the bot would not be able to manage any roles created before it joined the server.