Link to /join slash command when mentioned to user
Leverage the `</NAME:COMMAND_ID>` message formatting described at
https://discord.com/developers/docs/reference#message-formatting to link
to the `/join` slash command when it is referenced in a message from the
bot to the user.

To supports this functionality, add a cache of slash commands per guild
so that the `COMMAND_ID` can be quickly retrieved for any slash command
the bot implements.